+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added retract_url values to journals data
+- New class TheOJ::RetractionNotice to create/publish retracion notices and retract papers
 
 ## 1.6.0 (2023-08-08)